Move bcdofs to bcs.py
